requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
django-cors-headers>=4.0
django-environ>=0.11
django-ratelimit>=4.1
Pillow>=10.0
numpy>=1.21
opencv-python>=4.5
ultralytics>=8.0
torch>=2.0
